Create folders for a harmonization project

Creates folders for a harmonization project with a structure following the Maelstrom Research harmonization standards. This is an optional feature that can be used as a template for a harmonization project. The project can have versions or not. A new version can be created by copying from an existing version.

harmo_env_create(project, version = NA_character_, copy_from = NA_character_)

Arguments

project

A character string specifying the project name.

version

A character string specifying the project version.

copy_from

A character string specifying the existing project version to copy from.

Value

A project folder with eight subfolders and a README.Rmd file for documentation. If a version is specified, a subfolder for each version is created in the project folder, and the version folder contains the eight project subfolders and README file. If the project already exists, it will not be overwritten, but any missing subfolders will be created.

Examples

if (FALSE) {
# use case 1: create folder without versioning
harmo_env_create(project = "TEST_1")

# use case 2: create folder with a versioning
harmo_env_create(
  project = "TEST_2",
  version = "version_1")

# use case 3: copy a folder in the same project
# (e.g., the folder “version 2” will be created as a copy of the folder “version_1”)

harmo_env_create(
  project = "TEST_2",
  version = "version_2" ,
  copy_from = "version_1")
}